requirements.txt
flask
transformers
torch
python-docx
pypdf
reportlab
openai
